roundoff bugs corrected

+++ include/classes/gates.h
@@ -143,7 +143,7 @@ public:
 #pragma omp parallel for collapse(2)
         for (std::size_t j = 0; j < D; ++j)
             for (std::size_t i = 0; i < D; ++i)
-                result(i, j) = 1 / std::sqrt((double) D)
+                result(i, j) = 1 / std::sqrt(static_cast<double>(D))
                         * std::pow(omega(D), i * j);
 
         return result;
@@ -268,9 +268,10 @@ public:
         std::size_t ngate = subsys.size();
         std::size_t nctrl = ctrl.size();
         std::size_t nsubsys_bar = n - ctrlgate.size();
-        std::size_t D = std::llround(std::pow(d, n));
+        std::size_t D = static_cast<std::size_t>(std::llround(std::pow(d, n)));
         std::size_t DA = static_cast<std::size_t>(rA.rows());
-        std::size_t Dsubsys_bar = std::llround(std::pow(d, nsubsys_bar));
+        std::size_t Dsubsys_bar = static_cast<std::size_t>(
+                std::llround(std::pow(d, nsubsys_bar)));
 
         // compute the complementary subsystem of ctrlgate w.r.t. dims
         std::vector<std::size_t> allsubsys(n); // all subsystems
