+++ include/internal/functions.h
@@ -40,8 +40,8 @@ inline void _n2multiidx(std::size_t n, std::size_t numdims,
     // no error checks to improve speed
     for (std::size_t i = 0; i < numdims; ++i)
     {
-        result[numdims - i - 1] = n % static_cast<int>(dims[numdims - i - 1]);
-        n /= static_cast<int>(dims[numdims - i - 1]);
+        result[numdims - i - 1] = n % (dims[numdims - i - 1]);
+        n /= (dims[numdims - i - 1]);
     }
 }
